requirements.txt
flask
uvicorn[gunicorn]
python-dotenv
requests